requirements.txt
python-telegram-bot==20.7
groq==0.4.2
PyPDF2==3.0.1
